+++ test/test_replication.cpp
@@ -25,6 +25,7 @@
 #include <realm.hpp>
 #include <realm/util/features.h>
 #include <realm/util/file.hpp>
+#include <realm/util/overload.hpp>
 #include <realm/replication.hpp>
 
 #include "test.hpp"
@@ -264,55 +265,204 @@ TEST(Replication_WriteWithoutHistory)
     }
 }
 
-struct ObjectMutationObserver : _impl::NoOpTransactionLogParser {
-    unit_test::TestContext& test_context;
-    std::set<std::pair<TableKey, ObjKey>> expected_creations;
-    std::set<std::tuple<TableKey, ObjKey, ColKey>> expected_modifications;
+struct Select {
+    TableKey table_key;
+};
 
-    ObjectMutationObserver(unit_test::TestContext& test_context,
-                           std::initializer_list<std::pair<TableKey, int64_t>> creations,
-                           std::initializer_list<std::tuple<TableKey, int64_t, ColKey>> modifications)
+struct Create {
+    int64_t obj_key;
+};
+
+struct Mutate {
+    int64_t obj_key;
+    ColKey col_key;
+};
+
+struct Remove {
+    int64_t obj_key;
+};
+
+struct SelectColl {
+    int64_t obj_key;
+    ColKey col_key;
+};
+
+struct CollInsert {
+    size_t ndx;
+};
+
+struct CollSet {
+    size_t ndx;
+};
+
+using InstructionVariant = mpark::variant<Select, Create, Mutate, Remove, SelectColl, CollInsert, CollSet>;
+
+std::ostream& print_instructions(std::ostream& os, const std::vector<InstructionVariant>& ivs,
+                                 size_t first_difference) noexcept
+{
+    size_t ndx = 0;
+    for (auto& element : ivs) {
+        if (first_difference == ndx) {
+            os << "==> ";
+        }
+        util::format(os, "[%1]: ", ndx++);
+        auto print = overload{
+            [&](Select st) {
+                util::format(os, "Select{%1}", st.table_key);
+            },
+            [&](Create co) {
+                util::format(os, "CreateObject{%1}", co.obj_key);
+            },
+            [&](Mutate mo) {
+                util::format(os, "Mutate{%1, %2}", mo.obj_key, mo.col_key);
+            },
+            [&](Remove rm) {
+                util::format(os, "RemoveObject{%1}", rm.obj_key);
+            },
+            [&](SelectColl sc) {
+                util::format(os, "SelectCollection{%1, %2}", sc.obj_key, sc.col_key);
+            },
+            [&](CollInsert ci) {
+                util::format(os, "CollectionInsert{%1}", ci.ndx);
+            },
+            [&](CollSet cs) {
+                util::format(os, "CollectionSet{%1}", cs.ndx);
+            },
+        };
+        mpark::visit(print, element);
+        os << '\n';
+    }
+    return os;
+}
+
+bool compare_instructions(const InstructionVariant& a, const InstructionVariant& b)
+{
+    bool equal = false;
+    auto comp = overload{
+        [&](Select a_val) {
+            if (const Select* b_val = mpark::get_if<Select>(&b)) {
+                equal = a_val.table_key == b_val->table_key;
+            }
+        },
+        [&](Create a_val) {
+            if (const Create* b_val = mpark::get_if<Create>(&b)) {
+                equal = a_val.obj_key == b_val->obj_key;
+            }
+        },
+        [&](Mutate a_val) {
+            if (const Mutate* b_val = mpark::get_if<Mutate>(&b)) {
+                equal = (a_val.obj_key == b_val->obj_key && a_val.col_key == b_val->col_key);
+            }
+        },
+        [&](Remove a_val) {
+            if (const Remove* b_val = mpark::get_if<Remove>(&b)) {
+                equal = a_val.obj_key == b_val->obj_key;
+            }
+        },
+        [&](SelectColl a_val) {
+            if (const SelectColl* b_val = mpark::get_if<SelectColl>(&b)) {
+                equal = a_val.obj_key == b_val->obj_key && a_val.col_key == b_val->col_key;
+            }
+        },
+        [&](CollInsert a_val) {
+            if (const CollInsert* b_val = mpark::get_if<CollInsert>(&b)) {
+                equal = a_val.ndx == b_val->ndx;
+            }
+        },
+        [&](CollSet a_val) {
+            if (const CollSet* b_val = mpark::get_if<CollSet>(&b)) {
+                equal = a_val.ndx == b_val->ndx;
+            }
+        },
+    };
+    mpark::visit(comp, a);
+    return equal;
+}
+
+struct RecordingObserver : _impl::NoOpTransactionLogParser {
+    unit_test::TestContext& test_context;
+    std::vector<InstructionVariant> m_expected_ops;
+    std::vector<InstructionVariant> m_observed_ops;
+
+    RecordingObserver(unit_test::TestContext& test_context, std::initializer_list<InstructionVariant> ops)
         : test_context(test_context)
+        , m_expected_ops(ops.begin(), ops.end())
     {
-        for (auto [tk, ok] : creations) {
-            expected_creations.emplace(tk, ObjKey(ok));
-        }
-        for (auto [tk, ok, ck] : modifications) {
-            expected_modifications.emplace(tk, ObjKey(ok), ck);
-        }
     }
 
-    ObjectMutationObserver& operator=(ObjectMutationObserver&& obs) noexcept
+    RecordingObserver& operator=(RecordingObserver&& obs) noexcept
     {
-        expected_creations.swap(obs.expected_creations);
-        expected_modifications.swap(obs.expected_modifications);
+        m_observed_ops.swap(obs.m_observed_ops);
+        m_expected_ops.swap(obs.m_expected_ops);
         return *this;
+    }
+
+    bool select_table(TableKey t)
+    {
+        _impl::NoOpTransactionLogParser::select_table(t);
+        m_observed_ops.push_back(Select{t});
+        return true;
     }
 
     bool create_object(ObjKey obj_key)
     {
-        CHECK(expected_creations.erase(std::pair(get_current_table(), obj_key)));
+        m_observed_ops.push_back(Create{obj_key.value});
         return true;
     }
     bool modify_object(ColKey col, ObjKey obj)
     {
-        CHECK(expected_modifications.erase(std::tuple(get_current_table(), obj, col)));
+        m_observed_ops.push_back(Mutate{obj.value, col});
         return true;
     }
-    bool remove_object(ObjKey)
+    bool remove_object(ObjKey obj)
     {
+        m_observed_ops.push_back(Remove{obj.value});
+        return true;
+    }
+    bool select_collection(ColKey col_key, ObjKey obj_key, const StablePath& path)
+    {
+        _impl::NoOpTransactionLogParser::select_collection(col_key, obj_key, path);
+        m_observed_ops.push_back(SelectColl{obj_key.value, col_key});
+        return true;
+    }
+    bool collection_insert(size_t ndx)
+    {
+        m_observed_ops.push_back(CollInsert{ndx});
+        return true;
+    }
+    bool collection_set(size_t ndx)
+    {
+        m_observed_ops.push_back(CollSet{ndx});
         return true;
     }
 
     void check()
     {
-        CHECK(expected_creations.empty());
-        CHECK(expected_modifications.empty());
+        bool equality = m_observed_ops.size() == m_expected_ops.size();
+        size_t first_difference = -1;
+        if (equality) {
+            for (size_t i = 0; i < m_expected_ops.size(); ++i) {
+                if (!compare_instructions(m_observed_ops[i], m_expected_ops[i])) {
+                    first_difference = i;
+                    equality = false;
+                    break;
+                }
+            }
+        }
+
+        CHECK(equality);
+        if (!equality) {
+            std::cerr << "expected: \n";
+            print_instructions(std::cerr, m_expected_ops, first_difference);
+            std::cerr << "\nactual: \n";
+            print_instructions(std::cerr, m_observed_ops, first_difference);
+            std::cerr << std::endl;
+        }
     }
 };
 
 template <typename Fn>
-void expect(DBRef db, ObjectMutationObserver& observer, Fn&& write)
+void expect(DBRef db, RecordingObserver& observer, Fn&& write)
 {
     auto read = db->start_read();
     {
@@ -340,7 +490,7 @@ TEST(Replication_MutationsOnNewlyCreatedObject)
     }
 
     // Object creations with immediate mutations should report creations only
-    auto obs = ObjectMutationObserver(test_context, {{tk, 0}, {tk, 1}}, {});
+    auto obs = RecordingObserver(test_context, {Select{tk}, Create{0}, Create{1}});
     expect(db, obs, [](auto& tr) {
         auto table = tr.get_table("table");
         table->create_object().set_all(1);
@@ -348,7 +498,7 @@ TEST(Replication_MutationsOnNewlyCreatedObject)
     });
 
     // Mutating existing objects should report modifications
-    obs = ObjectMutationObserver(test_context, {}, {{tk, 0, col}, {tk, 1, col}});
+    obs = RecordingObserver(test_context, {Select{tk}, Mutate{0, col}, Mutate{1, col}});
     expect(db, obs, [](auto& tr) {
         auto table = tr.get_table("table");
         table->get_object(0).set_all(1);
@@ -358,7 +508,7 @@ TEST(Replication_MutationsOnNewlyCreatedObject)
     // Create two objects and then mutate them. We only track the most recently
     // created object, so this emits a mutation for the first object but not
     // the second.
-    obs = ObjectMutationObserver(test_context, {{tk, 2}, {tk, 3}}, {{tk, 2, col}});
+    obs = RecordingObserver(test_context, {Select{tk}, Create{2}, Create{3}, Mutate{2, col}});
     expect(db, obs, [](auto& tr) {
         auto table = tr.get_table("table");
         auto obj1 = table->create_object();
@@ -379,7 +529,7 @@ TEST(Replication_MutationsOnNewlyCreatedObject)
 
     // Creating an object in one table and then modifying the object with the
     // same ObjKey in a different table
-    obs = ObjectMutationObserver(test_context, {{tk2, 0}}, {{tk, 0, col}});
+    obs = RecordingObserver(test_context, {Select{tk2}, Create{0}, Select{tk}, Mutate{0, col}});
     expect(db, obs, [&](auto& tr) {
         auto table1 = tr.get_table(tk);
         auto table2 = tr.get_table(tk2);
@@ -393,7 +543,7 @@ TEST(Replication_MutationsOnNewlyCreatedObject)
     // Mutating an object whose Table has an index in group greater than the
     // higest of any created object after creating an object, which has to clear
     // the is-new-object flag
-    obs = ObjectMutationObserver(test_context, {{tk, 4}}, {{tk2, 0, col2}});
+    obs = RecordingObserver(test_context, {Select{tk}, Create{4}, Select{tk2}, Mutate{0, col2}});
     expect(db, obs, [&](auto& tr) {
         auto table1 = tr.get_table(tk);
         auto table2 = tr.get_table(tk2);
@@ -405,15 +555,15 @@ TEST(Replication_MutationsOnNewlyCreatedObject)
 
     // Splitting object creation and mutation over two different writes with the
     // same transaction object should produce mutation instructions
-    obs = ObjectMutationObserver(test_context, {{tk, 5}}, {{tk, 5, col}});
+    obs = RecordingObserver(test_context, {Select{tk}, Create{5}, Select{tk}, Mutate{5, col}});
     {
         auto read = db->start_read();
         auto tr = db->start_write();
         auto table = tr->get_table(tk);
-        auto obj = table->create_object();
+        auto obj = table->create_object(); // select tk
         tr->commit_and_continue_as_read();
         tr->promote_to_write();
-        obj.set_all(1);
+        obj.set_all(1); // select tk
         tr->commit_and_continue_as_read();
         read->advance_read(&obs);
         obs.check();
@@ -443,18 +593,18 @@ TEST(Replication_MutationsOnNewlyCreatedObject_Link)
     // Each top-level object creation is reported along with the mutation on
     // target_1 due to that both target objects are created before the mutations.
     // Nothing is reported for embedded objects
-    auto obs = ObjectMutationObserver(test_context, {{tk, 0}, {tk_target, 0}, {tk_target, 1}},
-                                      {{tk_target, 0, ck_target_value}});
+    auto obs = RecordingObserver(
+        test_context, {Select{tk}, Create{0}, Select{tk_target}, Create{0}, Create{1}, Mutate{0, ck_target_value}});
     expect(db, obs, [&](auto& tr) {
         auto table = tr.get_table(tk);
         auto target_table = tr.get_table(tk_target);
-        Obj obj = table->create_object();
-        Obj target_1 = target_table->create_object();
+        Obj obj = table->create_object();             // select tk
+        Obj target_1 = target_table->create_object(); // select tk_target
         Obj target_2 = target_table->create_object();
 
-        obj.set(ck_link_1, target_1.get_key());
+        obj.set(ck_link_1, target_1.get_key()); // select tk
         obj.set(ck_link_2, target_2.get_key());
-        target_1.set_all(1);
+        target_1.set_all(1); // select tk_target
         target_2.set_all(1);
 
         obj.create_and_set_linked_object(ck_embedded_1).set_all(1);
@@ -463,7 +613,8 @@ TEST(Replication_MutationsOnNewlyCreatedObject_Link)
 
     // Nullifying links via object deletions in both new and pre-existing objects
     // only reports the mutation in the pre-existing object
-    obs = ObjectMutationObserver(test_context, {{tk, 1}}, {{tk, 0, ck_link_1}});
+    obs =
+        RecordingObserver(test_context, {Select{tk}, Create{1}, Mutate{0, ck_link_1}, Select{tk_target}, Remove{0}});
     expect(db, obs, [&](auto& tr) {
         auto table = tr.get_table(tk);
         auto target_table = tr.get_table(tk_target);
@@ -502,7 +653,7 @@ TEST(Replication_MutationsOnNewlyCreatedObject_Collections)
 
     tr->commit();
 
-    auto obs = ObjectMutationObserver(test_context, {{tk, 0}, {tk_target, 0}}, {});
+    auto obs = RecordingObserver(test_context, {Select{tk}, Create{0}, Select{tk_target}, Create{0}});
     expect(db, obs, [&](auto& tr) {
         // Should report object creation but none of these mutations
         auto table = tr.get_table(tk);
@@ -526,6 +677,244 @@ TEST(Replication_MutationsOnNewlyCreatedObject_Collections)
         // mutations are on the newest object for each table
         obj.get_linklist(ck_embedded_list).create_and_insert_linked_object(0).set(ck_embedded_value, 1);
         obj.get_dictionary(ck_embedded_dictionary).create_and_insert_linked_object("a").set(ck_embedded_value, 1);
+    });
+
+    obs = RecordingObserver(test_context, {Select{tk},
+                                           SelectColl{0, ck_value_set},
+                                           CollInsert{1},
+                                           SelectColl{0, ck_value_list},
+                                           CollInsert{1},
+                                           CollInsert{2},
+                                           SelectColl{0, ck_value_dictionary},
+                                           CollSet{0},
+                                           CollInsert{1},
+                                           Select{tk_target},
+                                           Create{1},
+                                           Select{tk},
+                                           SelectColl{0, ck_obj_set},
+                                           CollInsert{1},
+                                           SelectColl{0, ck_obj_list},
+                                           CollInsert{1},
+                                           SelectColl{0, ck_obj_dictionary},
+                                           CollSet{0},
+                                           SelectColl{0, ck_embedded_list},
+                                           CollInsert{0},
+                                           SelectColl{0, ck_embedded_dictionary},
+                                           CollInsert{1}});
+    expect(db, obs, [&](auto& tr) {
+        // Should report mutations on this existing object
+        auto table = tr.get_table(tk);
+        Obj obj = table->get_object(0);
+        obj.get_set<int64_t>(ck_value_set).insert(5);
+        obj.get_list<int64_t>(ck_value_list).add(1);
+        obj.get_list<int64_t>(ck_value_list).add(2);
+        obj.get_dictionary(ck_value_dictionary).insert("a", 1);
+        obj.get_dictionary(ck_value_dictionary).insert("b", 2);
+
+        // Should report the object creation and the mutations on each collection
+        auto target_table = tr.get_table(tk_target);
+        Obj target_obj = target_table->create_object();
+        target_obj.set<int64_t>(ck_target_value, 2); // mutation skipped for this new object
+        obj.get_linkset(ck_obj_set).insert(target_obj.get_key());
+        obj.get_linklist(ck_obj_list).add(target_obj.get_key());
+        obj.get_dictionary(ck_obj_dictionary).insert("a", target_obj.get_key());
+
+        // Should not produce any instructions for the embedded objects created,
+        // just mutations on the obj which is not newly created.
+        obj.get_linklist(ck_embedded_list).create_and_insert_linked_object(0).set(ck_embedded_value, 1);
+        obj.get_dictionary(ck_embedded_dictionary).create_and_insert_linked_object("b").set(ck_embedded_value, 1);
+    });
+}
+
+TEST(Replication_NoSelectTableOnEmbeddedObjectMutations)
+{
+    SHARED_GROUP_TEST_PATH(path);
+    auto db = DB::create(make_in_realm_history(), path);
+    auto tr = db->start_write();
+
+    auto table = tr->add_table("table");
+    auto tk = table->get_key();
+    ColKey ck_value = table->add_column(type_Int, "value");
+
+    auto target_table = tr->add_table("target table");
+    auto tk_target = target_table->get_key();
+    auto ck_target_value = target_table->add_column(type_Int, "value");
+
+    auto embedded_table = tr->add_table("embedded table", Table::Type::Embedded);
+    auto tk_embedded = embedded_table->get_key();
+    auto ck_embedded_value = embedded_table->add_column(type_Int, "value");
+    auto embedded_table2 = tr->add_table("embedded_table2", Table::Type::Embedded);
+    auto tk_embedded2 = embedded_table2->get_key();
+    auto ck_embedded2_str = embedded_table2->add_column(type_String, "value_str");
+    auto ck_embedded_link = embedded_table->add_column(*embedded_table2, "embed");
+    ColKey ck_embedded_dictionary = table->add_column_dictionary(*embedded_table, "embedded dictionary");
+
+    tr->commit();
+
+    auto obs = RecordingObserver(test_context, {Select{tk}, Create{0}, Select{tk_target}, Create{0}, Create{1}});
+    expect(db, obs, [&](auto& tr) {
+        // Should report object creation but none of these mutations
+        auto table = tr.get_table(tk);
+        Obj obj = table->create_object();
+        obj.set<int64_t>(ck_value, 1);
+
+        // Should report the object creation but not the mutations on either object,
+        // as they're both the most recently created object in each table
+        auto target_table = tr.get_table(tk_target);
+        Obj target_obj = target_table->create_object(); // select tk_target
+        target_obj.set<int64_t>(ck_target_value, 1);
+
+        // Should not produce any instructions: embedded object creations aren't
+        // replicated (as you can't observe embedded tables directly), and the
+        // mutations are on the newest object for each table
+        Obj embedded_a = obj.get_dictionary(ck_embedded_dictionary).create_and_insert_linked_object("a");
+        embedded_a.set(ck_embedded_value, 1);
+        Obj embedded2_a = embedded_a.create_and_set_linked_object(ck_embedded_link);
+        embedded2_a.set(ck_embedded2_str, "test");
+
+        // target_table should still be selected, because there should have been
+        // no select table emitted for the embedded objects above
+        Obj target_obj2 = target_table->create_object();
+        target_obj2.set<int64_t>(ck_target_value, 2);
+    });
+
+    obs = RecordingObserver(test_context, {Select{tk}, Create{1}, Select{tk_embedded}, Mutate{1, ck_embedded_value},
+                                           Select{tk_embedded2}, Mutate{1, ck_embedded2_str}});
+    expect(db, obs, [&](auto& tr) {
+        // Should report object creation but none of these mutations
+        auto table = tr.get_table(tk);
+        Obj obj = table->create_object(); // select tk
+        obj.set<int64_t>(ck_value, 1);
+
+        // Should not produce any instructions: embedded object creations aren't
+        // replicated (as you can't observe embedded tables directly), and the
+        // mutations are on the newest object for each table
+        Obj embedded_a = obj.get_dictionary(ck_embedded_dictionary).create_and_insert_linked_object("a");
+        embedded_a.set(ck_embedded_value, 1);
+        Obj embedded2_a = embedded_a.create_and_set_linked_object(ck_embedded_link);
+        embedded2_a.set(ck_embedded2_str, "test a");
+
+        Obj embedded_b = obj.get_dictionary(ck_embedded_dictionary).create_and_insert_linked_object("b");
+        embedded_b.set(ck_embedded_value, 2);
+        Obj embedded2_b = embedded_b.create_and_set_linked_object(ck_embedded_link);
+        embedded2_b.set(ck_embedded2_str, "test b");
+
+        // setting a property on embeded_a means that it is no longer the newest object
+        // created, so we require a set_table, and modification for each modify:
+        // select "embedded table"
+        // modify embedded_a
+        // select "embedded table 2"
+        // modify embedded2_a
+        embedded_a.set(ck_embedded_value, 3);
+        embedded2_a.set(ck_embedded2_str, "test a 2");
+    });
+}
+
+TEST(Replication_EmbeddedListInsertions)
+{
+    SHARED_GROUP_TEST_PATH(path);
+    auto db = DB::create(make_in_realm_history(), path);
+    auto tr = db->start_write();
+
+    auto table = tr->add_table("table");
+    auto tk = table->get_key();
+    ColKey ck_value = table->add_column(type_Int, "value");
+
+    auto embedded_table = tr->add_table("embedded table", Table::Type::Embedded);
+    auto tk_embedded = embedded_table->get_key();
+    auto ck_embedded_value = embedded_table->add_column(type_Int, "value");
+    ColKey ck_embedded_list = table->add_column_list(*embedded_table, "embedded list");
+
+    // initial state
+    int64_t obj_key;
+    {
+        auto table = tr->get_table(tk);
+        Obj obj = table->create_object();
+        obj.set<int64_t>(ck_value, 1);
+        obj_key = obj.get_key().value;
+    }
+
+    tr->commit();
+
+    auto obs = RecordingObserver(test_context, {Select{tk}, SelectColl{obj_key, ck_embedded_list}, CollInsert{0},
+                                                CollInsert{1}, Select{tk_embedded}, Mutate{0, ck_embedded_value}});
+    expect(db, obs, [&](auto& tr) {
+        auto table = tr.get_table(tk);
+        Obj obj = table->get_object(0);
+        LnkLst list = obj.get_linklist(ck_embedded_list);
+        Obj link_0 = list.create_and_insert_linked_object(0);
+        Obj link_1 = list.create_and_insert_linked_object(1);
+
+        // both of these were just created, but only one is the "recently created"
+        // so we do record one unnecessary mutation
+        link_0.set<int64_t>(ck_embedded_value, 10);
+        link_1.set<int64_t>(ck_embedded_value, 11);
+    });
+}
+
+TEST(Replication_EmbeddedListInsertionsWithListMutations)
+{
+    SHARED_GROUP_TEST_PATH(path);
+    auto db = DB::create(make_in_realm_history(), path);
+    auto tr = db->start_write();
+
+    auto table = tr->add_table("table");
+    auto tk = table->get_key();
+    ColKey ck_value = table->add_column(type_Int, "value");
+
+    auto embedded_table = tr->add_table("embedded table", Table::Type::Embedded);
+    auto tk_embedded = embedded_table->get_key();
+    ColKey ck_embedded_list_of_ints = embedded_table->add_column_list(type_Int, "int list");
+    ColKey ck_list_of_embeddeds = table->add_column_list(*embedded_table, "embedded list");
+
+    // initial state
+    int64_t obj_key;
+    {
+        auto table = tr->get_table(tk);
+        Obj obj = table->create_object();
+        obj.set<int64_t>(ck_value, 1);
+        obj_key = obj.get_key().value;
+    }
+
+    tr->commit();
+
+    // there should be no extra select collection on ck_list_of_embeddeds between insertions,
+    // even though there are modifications to the embedded lists between insertions
+    auto obs = RecordingObserver(
+        test_context, {Select{tk}, SelectColl{obj_key, ck_list_of_embeddeds}, CollInsert{0}, CollInsert{1}});
+    expect(db, obs, [&](auto& tr) {
+        auto table = tr.get_table(tk);
+        Obj obj = table->get_object(0);
+        LnkLst list = obj.get_linklist(ck_list_of_embeddeds);
+        Obj link_0 = list.create_and_insert_linked_object(0);
+        link_0.get_list<Int>(ck_embedded_list_of_ints).insert(0, 0);
+        link_0.get_list<Int>(ck_embedded_list_of_ints).insert(1, 1);
+        Obj link_1 = list.create_and_insert_linked_object(1);
+        link_1.get_list<Int>(ck_embedded_list_of_ints).insert(0, 10);
+        link_1.get_list<Int>(ck_embedded_list_of_ints).insert(1, 11);
+    });
+
+    // modifications on an existing embedded object should make selections on the collection
+    obs = RecordingObserver(test_context,
+                            {Select{tk_embedded}, SelectColl{0, ck_embedded_list_of_ints}, CollInsert{0}, Select{tk},
+                             SelectColl{obj_key, ck_list_of_embeddeds}, CollInsert{2}, Select{tk_embedded},
+                             SelectColl{1, ck_embedded_list_of_ints}, CollInsert{0}});
+    expect(db, obs, [&](auto& tr) {
+        auto table = tr.get_table(tk);
+        Obj obj = table->get_object(0);
+        LnkLst list = obj.get_linklist(ck_list_of_embeddeds);
+        CHECK_EQUAL(list.size(), 2);
+        Obj link_0 = list.get_object(0);
+        // select embedded table, select embedded collection, collection insert
+        link_0.get_list<Int>(ck_embedded_list_of_ints).insert(0, 100);
+
+        // select top table, select top collection, collection insert
+        Obj link_2 = list.create_and_insert_linked_object(2);
+        link_2.get_list<Int>(ck_embedded_list_of_ints).insert(0, 0);
+
+        // select embedded table, select embedded collection, collection insert
+        Obj link_1 = list.get_object(1);
+        link_1.get_list<Int>(ck_embedded_list_of_ints).insert(0, 1000);
     });
 }